Use a separate index for column -> property lookups rather than sorting the properties
This makes things a bit less weird when constructing a schema manually for the
dynamic API (since the properties array visible to you will actually match what
you passed in), is required to support having the array of properties be a
subset of the table columns, and simplifies handling columns being rearranged a bit.

Also a little faster since it's O(N) rather than O(N log N), but if you have
enough properties for it to be even remotely noticeable something has gone very
wrong.